+++ tests/test_utils.py
@@ -13,7 +13,7 @@ from stable_baselines3.common.env_util import is_wrapped, make_atari_env, make_v
 from stable_baselines3.common.evaluation import evaluate_policy
 from stable_baselines3.common.monitor import Monitor
 from stable_baselines3.common.noise import ActionNoise, OrnsteinUhlenbeckActionNoise, VectorizedActionNoise
-from stable_baselines3.common.utils import polyak_update, system_info, zip_strict
+from stable_baselines3.common.utils import get_system_info, polyak_update, zip_strict
 from stable_baselines3.common.vec_env import DummyVecEnv, SubprocVecEnv
 
 
@@ -379,8 +379,8 @@ def test_ppo_warnings():
         PPO("MlpPolicy", "Pendulum-v0", n_steps=6, batch_size=8)
 
 
-def test_system_info():
-    info, info_str = system_info(print_info=True)
+def test_get_system_info():
+    info, info_str = get_system_info(print_info=True)
     assert info["Stable-Baselines3"] == str(sb3.__version__)
     assert "Python" in info_str
     assert "PyTorch" in info_str